20250613_global_logic/gl_050_python_scripting/file_processor/pyproject.toml
[project]
name = "file-processor"
version = "0.1.0"
description = "File processing utility for DevOps operations"
authors = [{name = "DevOps Team", email = "devops@example.com"}]
dependencies = [
    "PyYAML>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"